Moving stuff to afterEvaluate
