+++ connectors/format/csv/build.gradle.kts
@@ -1,9 +1,35 @@
+plugins {
+    alias(libs.plugins.android)
+}
+
 kotlin {
+    linuxX64()
+    linuxArm64()
+
+    mingwX64()
+
+    js(IR) {
+        browser()
+        nodejs()
+    }
+
+    androidTarget {
+        publishLibraryVariants("release", "debug")
+    }
+
+    macosArm64()
+    macosX64()
+
+    iosArm64()
+
+    tvosArm64()
+
+    watchosArm64()
+
     sourceSets {
-        jvmMain {
+        commonMain {
             dependencies {
-                val modules = modules { project(it) }
-                api(modules.file)
+                api(libs.kotlin.reflect)
             }
         }
     }
