+++ core/src/commonMain/kotlin/com/river/core/SuspendExt.kt
@@ -1,7 +1,10 @@
 package com.river.core
 
-import kotlinx.coroutines.*
+import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.CoroutineStart.LAZY
+import kotlinx.coroutines.Deferred
+import kotlinx.coroutines.async
+import kotlinx.coroutines.coroutineScope
 
 /**
  * Extension property on a suspend function that allows it to be called asynchronously.
